+++ src/ecs/mod.rs
@@ -1,9 +1,11 @@
 //module tree
 mod callbacks;
 mod component_utils;
+mod react;
 mod system_callers;
 
 //API exports
 pub use crate::ecs::callbacks::*;
 pub use crate::ecs::component_utils::*;
+pub use crate::ecs::react::*;
 pub use crate::ecs::system_callers::*;