+++ tests/dummy_test.rs
@@ -1,4 +1,0 @@
-//todo: add a test
-#[test]
-fn placeholder()
-{}