+++ tests/tests.rs
@@ -1,2 +1,0 @@
-//module tree
-mod dummy_test;